+++ lib/led/led.h
@@ -11,6 +11,6 @@ void setup_led(void);
 void led_blink(void);
 void led_on(void);
 void led_off(void);
-void led_pwm_dimming(double brightness);
+void led_pwm_dimming(int brightness);
 
 #endif
